+++ karma.conf.js
@@ -16,6 +16,7 @@ module.exports = (config) => {
       'node_modules/font-awesome/css/font-awesome.css',
       'build/hglib.css',
       'test/**/*.+(js|jsx)',
+      // 'test/TrackLabelsTest.jsx',
       // 'test/AxisTests.js',
       // 'test/SVGExportTest.js',
       // 'test/AxisTests.js', // Passes when run in isolation, but fails as part of suite.