+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## v1.3.1
 
 - Add `constIndicators` as an option to 1D tracks for displaying a visual indicator (e.g., a line) at a constant value.
+- Added `[glyph-color]` as the default option for line and bar tracks
 
 ## v1.3.0